feat: AMQP queues as abstraction over queue names
