+++ src/mq/index.ts
@@ -1,5 +1,5 @@
 export * from "./amqp/manager";
-export { AMQPQueue } from "./amqp/queue";
+export { Queue } from "./amqp/queue";
 export { AMQPWorker, command, jobs } from "./amqp/worker";
 export { groupDecorator, handlerDecorator, ParsedHandler, parseHandlers } from "./decorators";
 export * from "./handlers";
